feat : API를 이용한 카카오 로그인

+++ src/main/java/com/green/service/ReplyService.java
@@ -8,9 +8,9 @@ import com.green.vo.ReplyVO;
 
 public interface ReplyService {
 	public int register(ReplyVO vo);
-	public ReplyVO get(Long rno);
+	public ReplyVO get(int rno);
 	public int modify(ReplyVO vo);
-	public int remove(Long rno);
-	public List<ReplyVO> getList(Criteria cri, Long bno);
-	public ReplyPageDTO getListPage(Criteria cri,Long bno);
+	public int remove(int rno);
+	public List<ReplyVO> getList(Criteria cri, int bno);
+	public ReplyPageDTO getListPage(Criteria cri,int bno);
 }
